feat: add shared layout for pages
